V2.4.12 Fixed a issue where you couldn't permanentely delete a reminder if there were no advanced reminder properties attached to the reminder

+++ Data Access Layer/DLAVRProperties.cs
@@ -112,9 +112,12 @@ namespace Data_Access_Layer
 
         public static void DeleteAvrProperties(long id)
         {
+            AdvancedReminderProperties prop = GetAVRProperties(id);
+            if (prop == null)
+                return;
+
             using (RemindMeDbEntities db = new RemindMeDbEntities())
-            {
-                AdvancedReminderProperties prop = GetAVRProperties(id);
+            {                
                 db.AdvancedReminderProperties.Attach(prop);
                 db.AdvancedReminderProperties.Remove(prop);
